Update Developer Guide Glossary
The glossary does not contain the terms related to trip planning, which
can result in confusion.

Writing the terms for each aspect of Maplet will allow for the group
to be on the same page when implementing features.

Explicitly stating the relationship between the aspects of Maplet allows
for clearer understanding of the features to be implemented.