Fix OpeningHours
OpeningHours will throw an error if the field was empty. which is
unintended behaviour.

Supporting a "null" version of OpeningHours will fix this problem.

This implementation will allow for it to be supported for user inputs,
and will still have the functionalities such as checking if an
attraction is open at a specific time.

The regex for 24 hour time is supplemented with examples of both valid
and invalid for better readability.