Fix Price-related bugs
The Price error messgae does not fully outline what the field can
actually acept.

Further restricting the inputs, as well as specifying what is allowed
for a price field consistent behaviour, and ensures a consistent format
when inputting prices into Maplet.

To be clearer with the user on what is and isn't allowed, and to
ensure that Maplet is usable and clear on what is an acceptable format.